// src/app/models/question.model.ts
export interface Question {
    category: string;
    type: string;
    difficulty: string;
    question: string;
    correct_answer: string;
    incorrect_answers: Array<string>,
    answers?: Answer[];
   
}

export interface Answer{
    description: string;
    isCorrect: boolean;
    selected: boolean;
}